+++ src/euroeval/dataset_configs/catalan.py
@@ -2,7 +2,7 @@
 
 from ..data_models import DatasetConfig
 from ..languages import CATALAN
-from ..tasks import COMMON_SENSE, KNOW, LA, NER, RC, SENT, SUMM
+from ..tasks import COMMON_SENSE, INSTRUCTION_FOLLOWING, KNOW, LA, NER, RC, SENT, SUMM
 
 # Official datasets ###
 
@@ -61,4 +61,14 @@ WINOGRANDE_CA_CONFIG = DatasetConfig(
     task=COMMON_SENSE,
     languages=[CATALAN],
     labels=["a", "b"],
+)
+
+IFEVAL_CA_CONFIG = DatasetConfig(
+    name="ifeval-ca",
+    pretty_name="IFEval-ca",
+    source="EuroEval/ifeval-ca",
+    task=INSTRUCTION_FOLLOWING,
+    languages=[CATALAN],
+    train_split=None,
+    val_split=None,
 )